+++ resale/src/main/java/com/goti/resale/service/domain/ResaleRestrictionDomainService.java
@@ -1,0 +1,9 @@
+package com.goti.resale.service.domain;
+
+import java.util.UUID;
+
+import com.goti.resale.domain.entity.resale.ResaleRestrictionEntity;
+
+public interface ResaleRestrictionDomainService {
+	ResaleRestrictionEntity getOrCreateRestriction(UUID userId);
+}
